Revert "設定保存とUI・サービス分離のリファクタリング"
