--- Packages/com.sunagimo.tilemapsplitter/Editor/TilemapSplitSettings.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 265316bbd1df4751ab7e2e10d847bf59
