--- Packages/com.sunagimo.tilemapsplitter/Editor/TilemapSplitSettingsRepository.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 1634adb7d2e548208c084c1b4b058688
