--- Packages/com.sunagimo.tilemapsplitter/Editor/TilemapSplitterService.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: f5338654ba3e4553affba7e289a36a62
