--- Packages/com.sunagimo.tilemapsplitter/Editor/TilemapSplitterUI.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 24e388d6677a45bd8279323dd258edff
